requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.29.0
pydantic>=2.10.0
psutil>=6.0.0
watchdog>=4.0.0

# Database
sqlalchemy>=2.0.30
aiosqlite>=0.20.0

# LLM & Threat Intel
anthropic>=0.40.0
# ollama>=0.2.0
stix2>=3.0.1

# Security & Parsing
python-evtx>=0.7.4

# Testing & Simulation
pytest>=8.2.0
pytest-asyncio>=0.23.6
httpx>=0.27.0

# Utilities
pyyaml>=6.0.1
python-dotenv>=1.0.1
websockets>=12.0
jinja2>=3.1.4
